refactor: change repositories parent folder name
