+++ Doppler.ImageAnalyzer.Api/Services/Repositories/Entities/ImageAnalysisResultDocumentInfo.cs
@@ -1,4 +1,4 @@
-﻿namespace Doppler.ImageAnalyzer.Api.Services.MongoDB.Collections
+﻿namespace Doppler.ImageAnalyzer.Api.Services.Respositories.Entities
 {
     public static class ImageAnalysisResultDocumentInfo
     {
